requirements.txt
pytest==5.3.5
uvicorn==0.12.3
fastapi==0.61.2
python-dateutil==2.8.2